requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
batch-processor==0.0.3
beautifulsoup4==4.13.5
boto3==1.40.29
botocore==1.40.29
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
colorthief==0.2.1
distro==1.9.0
docopt==0.6.2
fake-useragent==2.2.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.8.0
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
multidict==6.7.0
numpy==2.3.3
openai==1.107.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pipreqs==0.4.13
propcache==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
redis==7.1.0
reportlab==4.4.3
requests==2.32.5
requests-file==2.1.0
requests-ip-rotator==1.0.14
s3transfer==0.14.0
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.47.3
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarg==0.1.10
yarl==1.22.0
